% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge_params.R
\name{badge_params}
\alias{badge_params}
\title{Build \code{badge_params}}
\usage{
badge_params(value)
}
\arguments{
\item{value}{(numeric) coverage value}
}
\value{
a list (a \code{badge_params} object) with the following elements:
\itemize{
\item value_num: the coverage value as numeric (adjusted, if necessary)
\item value_char: the coverage value as character
\item value_col: the colour for the \emph{value} box
\item width_label: the width of the \emph{label} box
\item width_value: the width of the \emph{value} box
\item text_length_label: the length of the \emph{label} text
\item text_length_value: the length of the \emph{value} text
\item total_width: the total badge width
\item text_start_value: point along the x-axis where the \emph{value} text should
start
}
}
\description{
Take the value and build a list with all the value-dependent parameters
to be injected into the SVG template.
}
\keyword{internal}
