% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{build_badge_url}
\alias{build_badge_url}
\title{Build the URL to the badge SVG}
\usage{
build_badge_url(value)
}
\arguments{
\item{value}{(numeric) coverage value}
}
\value{
a glue string.
}
\description{
Build the URL to the badge SVG
}
\keyword{internal}
