% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_coverage.R
\name{calculate_file_coverage}
\alias{calculate_file_coverage}
\title{Calculate file coverage}
\usage{
calculate_file_coverage(x = covr::package_coverage())
}
\arguments{
\item{x}{\verb{<coverage>} object, defaults to \code{\link[covr:package_coverage]{covr::package_coverage()}}.}
}
\value{
a \code{tibble} with 2 columns (\code{File} and \code{Coverage}) summarising the
coverage at file level.
}
\description{
Calculate file coverage
}
\examples{
\dontrun{
library(covr)

covr::package_coverage("myawesomepkg") |>
  calculate_file_coverage()
}
}
