% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{classify_lines}
\alias{classify_lines}
\title{Classify lines}
\usage{
classify_lines(raw_diff_df)
}
\arguments{
\item{raw_diff_df}{(tibble) a raw diff tibble with a single column
named `raw}
}
\value{
a tibble with 5 additional columns, all logical, classifying the
rows into:
\itemize{
\item hunk:
\item base: present only in base (deleted)
\item head: present only in head (added)
\item merge: present neither in head nor in base
\item context: present both in head and in base
}
}
\description{
\code{-}: line appears in base, but not in head (was deleted)
\code{+}: line appears in head, but not in base (was added)
\verb{++}: line was added, but does not appear in either head or base
it indicates a line that did not exist in either parent and was
introduced by the merge resolution itself
\code{@}: hunk header
}
\keyword{internal}
