% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_coverage.R
\name{combine_file_coverage}
\alias{combine_file_coverage}
\title{Derive a coverage tibble for the changed files}
\usage{
combine_file_coverage(head_coverage, base_coverage)
}
\arguments{
\item{head_coverage}{(coverage) active / current branch (\code{HEAD}) coverage.
The output of \code{\link[covr:package_coverage]{covr::package_coverage()}} on the branch.}

\item{base_coverage}{(coverage) base / target branch coverage (coverage for
the branch merging into). The output of \code{\link[covr:package_coverage]{covr::package_coverage()}} on the
branch.}
}
\value{
a\code{tibble} with 4 columns:
\itemize{
\item \code{file}: file name
\item \code{coverage_head}: coverage for the current branch
\item \code{coverage_base} coverage for the base branch, and
\item \code{delta}: difference in coverage between head and base.
}
}
\description{
Derive a coverage tibble for the changed files
}
\keyword{internal}
