% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{compose_comment}
\alias{compose_comment}
\title{Compose a coverage comment}
\usage{
compose_comment(
  head_coverage,
  base_coverage,
  repo,
  pr_number,
  marker = "<!-- covr2md-code-coverage -->"
)
}
\arguments{
\item{head_coverage}{(coverage) active / current branch (\code{HEAD}) coverage.
The output of \code{\link[covr:package_coverage]{covr::package_coverage()}} on the branch.}

\item{base_coverage}{(coverage) base / target branch coverage (coverage for
the branch merging into). The output of \code{\link[covr:package_coverage]{covr::package_coverage()}} on the
branch.}

\item{repo}{(character) the repository name in the GitHub format
(\code{"OWNER/REPO"}).}

\item{pr_number}{(integer) the PR number}

\item{marker}{(character scalar) string used to identify an issue
comment generated with covr2md. Defaults to
\code{"<!-- covr2md-code-coverage -->"}.}
}
\value{
a character scalar with the content of the GitHub comment
}
\description{
Compose a coverage comment
}
\examples{
\dontrun{
head_coverage <- covr::package_coverage()
system("git checkout main")
base_coverage <- covr::package_coverage()

compose_comment(
  head_coverage = head_coverage,
  base_coverage = base_coverage,
  repo = "dragosmg/covr2mddemo",
  pr_number = 3
)
}
}
