% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{compose_coverage_details}
\alias{compose_coverage_details}
\title{Compose coverage details}
\usage{
compose_coverage_details(
  head_coverage,
  base_coverage,
  changed_files,
  keep_all_files = FALSE,
  align = "rrrc"
)
}
\arguments{
\item{head_coverage}{(coverage) active / current branch (\code{HEAD}) coverage.
The output of \code{\link[covr:package_coverage]{covr::package_coverage()}} on the branch.}

\item{base_coverage}{(coverage) base / target branch coverage (coverage for
the branch merging into). The output of \code{\link[covr:package_coverage]{covr::package_coverage()}} on the
branch.}

\item{changed_files}{(character) names of files changed by the PR. Usually
the output of \code{\link[=get_changed_files]{get_changed_files()}}.}

\item{keep_all_files}{(logical) include all files in the diff coverage table
or just those modified by the PR.}

\item{align}{Column alignment: a character vector consisting of \code{'l'}
(left), \code{'c'} (center) and/or \code{'r'} (right). By default or if
\code{align = NULL}, numeric columns are right-aligned, and other columns
are left-aligned. If \code{length(align) == 1L}, the string will be
expanded to a vector of individual letters, e.g. \code{'clc'} becomes
\code{c('c', 'l', 'c')}, unless the output format is LaTeX.}
}
\value{
a markdown table of changes in coverage at file level between the
head and base branches.
}
\description{
Compose coverage details
}
\examples{
\dontrun{
changed_files <- get_changed_files(
  repo = "dragosmg/covr2mddemo",
  pr_number = 2
)

coverage_details <- compose_coverage_details(
  head_coverage = head_coverage,
  base_coverage = base_coverage,
  changed_files = changed_files
)
}
}
\keyword{internal}
