% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{compose_coverage_summary}
\alias{compose_coverage_summary}
\title{Compose coverage summary}
\usage{
compose_coverage_summary(pr_details, delta)
}
\arguments{
\item{pr_details}{a \verb{<pr_details>} object representing a subset of the pull
request metadata we need. The output of \code{\link[=get_pr_details]{get_pr_details()}}.}

\item{delta}{(numeric scalar) difference in total coverage between head of
current branch and base branch.}
}
\value{
a string (character scalar) containing the text for the coverage
summary.
}
\description{
Builds the top level sentences of the summary.
}
\examples{
\dontrun{
pr_details <- get_pr_details(
  repo = "<owner>/<repo>",
  pr_number = 2
)

coverage_summary <- compose_coverage_summary(
  pr_details = pr_details,
  delta = 20.43
)
}
}
\keyword{internal}
