% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_coverage.R
\name{compose_file_coverage_details}
\alias{compose_file_coverage_details}
\title{Compose the file coverage details section}
\usage{
compose_file_coverage_details(file_cov_df)
}
\arguments{
\item{file_cov_df}{a \code{tibble}, the output of \code{\link[=combine_file_coverage]{combine_file_coverage()}}}
}
\value{
a glue object, a string with the section content.
}
\description{
the section is made up of a subtitle (heading 3) and a table. if the input is
\code{NULL}, the output is an empty string.
}
\keyword{internal}
