% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_coverage.R
\name{compose_line_coverage_details}
\alias{compose_line_coverage_details}
\title{Compose the line coverage details section}
\usage{
compose_line_coverage_details(diff_line_coverage)
}
\arguments{
\item{diff_line_coverage}{a \code{tibble} the output of \code{\link[=get_diff_line_coverage]{get_diff_line_coverage()}}}
}
\value{
a \code{glue} string
}
\description{
The section is made up of a subtitle (Heading 3) and a Markdown table. If
the input object is \code{NULL} an empty string is returned.
}
\details{
This section will be part of the "More details" collapsible section of the
GitHub comment.
}
\keyword{internal}
