% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_coverage.R
\name{compose_line_coverage_summary}
\alias{compose_line_coverage_summary}
\title{Compose the line coverage summary}
\usage{
compose_line_coverage_summary(diff_line_coverage, target = 80)
}
\arguments{
\item{diff_line_coverage}{a \code{tibble} the output of \code{\link[=get_diff_line_coverage]{get_diff_line_coverage()}}}

\item{target}{(numeric) the target coverage for the diff. Default to 80, but
most often the base total coverage should be used.}
}
\value{
a glue string, a sentence summarising the diff coverage.
}
\description{
Builds the high-level sentence summarising the line coverage of the patch.
}
\keyword{internal}
