% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covr2gh-package.R
\docType{package}
\name{covr2gh-package}
\alias{covr2gh}
\alias{covr2gh-package}
\title{covr2gh: Use covr Output in GitHub}
\description{
Post code coverage insights as GitHub comments.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dragosmg/covr2gh}
  \item \url{https://dragosmg.github.io/covr2gh/}
  \item Report bugs at \url{https://github.com/dragosmg/covr2gh/issues}
}

}
\author{
\strong{Maintainer}: Dragoș Moldovan-Grünfeld \email{dragos.mold@gmail.com}

}
\keyword{internal}
