% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covr2md-package.R
\docType{package}
\name{covr2md-package}
\alias{covr2md}
\alias{covr2md-package}
\title{covr2md: Tools to Convert covr Output to Markdown}
\description{
Convert covr code coverage output into Markdown in preparation for use with GitHub comments.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dragosmg/covr2md}
  \item Report bugs at \url{https://github.com/dragosmg/covr2md/issues}
}

}
\author{
\strong{Maintainer}: Dragoș Moldovan-Grünfeld \email{dragos.mold@gmail.com}

}
\keyword{internal}
