% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment.R
\name{delete_comment}
\alias{delete_comment}
\title{Delete a comment}
\usage{
delete_comment(repo, comment_id)
}
\arguments{
\item{repo}{(character) the repository name in the GitHub format
(\code{"OWNER/REPO"}).}

\item{comment_id}{(numeric) the ID of the issue comment to delete.}
}
\value{
a \code{gh_response} object
}
\description{
Thin wrapper for making a \code{DELETE} request to the GitHub API.
}
\examples{
\dontrun{
delete_comment("dragosmg/this-is-my-repo", 4553)
}
}
\keyword{internal}
