% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge_params.R
\name{derive_badge_colour}
\alias{derive_badge_colour}
\title{Derive the colour for the value part of the badge}
\usage{
derive_badge_colour(badge_value, colours = coverage_thresholds)
}
\arguments{
\item{colours}{(\code{tibble}) a tibble with 2 columns \code{value} (threshold) and
\code{colour}.}

\item{value}{(integer-like scalar) coverage percentage}
}
\value{
a colour hexcode as string
}
\description{
Maps a value to a certain interval and chooses the corresponding colour.
}
\examples{
\dontrun{
derive_badge_colour(5)
}
}
\keyword{internal}
