% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/digest.R
\name{derive_diff_df}
\alias{derive_diff_df}
\title{Derive a coverage tibble for the changed files}
\usage{
derive_diff_df(
  head_coverage,
  base_coverage,
  changed_files,
  keep_all_files = FALSE,
  keep_totals = TRUE
)
}
\arguments{
\item{head_coverage}{(coverage) active / current branch (\code{HEAD}) coverage.
The output of \code{\link[covr:package_coverage]{covr::package_coverage()}} on the branch.}

\item{base_coverage}{(coverage) base / target branch coverage (coverage for
the branch merging into). The output of \code{\link[covr:package_coverage]{covr::package_coverage()}} on the
branch.}

\item{keep_all_files}{(logical) indicates whether to subset only for the
relevant (i.e. under \verb{R/} or \verb{src/}) changed files. Defaults to \code{FALSE}.}
}
\value{
a\code{tibble} with 4 columns:
\itemize{
\item \code{file}: file name
\item \code{coverage_head}: coverage for the current branch
\item \code{coverage_base} coverage for the base branch, and
\item \code{delta}: difference in coverage between head and base.
}
}
\description{
Derive a coverage tibble for the changed files
}
\keyword{internal}
