% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff.R
\name{diff_split}
\alias{diff_split}
\title{Split (un-combine) diff text}
\usage{
diff_split(diff_text)
}
\arguments{
\item{diff_text}{(character) string with the diff contents}
}
\value{
a list with 2 tibbles:
\itemize{
\item \code{head_lines}: the numbered lines that were added in head
\item \code{base_lines}: the numbered lines that were removed from base
}

Each \code{tibble} has 2 columns: \code{line} and `source``
}
\description{
Reverse-engineer the combined view of the diff.
}
\keyword{internal}
