% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge_params.R
\name{estimate_text_length_value}
\alias{estimate_text_length_value}
\title{Estimate the length of the value text}
\usage{
estimate_text_length_value(badge_value, text_length_label = 50)
}
\arguments{
\item{badge_value}{(\code{badge_value}) the output of \code{badge_value()}}

\item{text_length_label}{(numeric) the length of the label text. Defaults to
50 pixels.}
}
\value{
numeric representing the length of the label text
}
\description{
Estimate the length of the value text
}
\keyword{internal}
