% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge_params.R
\name{estimate_width_value}
\alias{estimate_width_value}
\title{Estimate the width of the value box}
\usage{
estimate_width_value(badge_value, width_label = 60)
}
\arguments{
\item{badge_value}{(\code{badge_value}) the output of \code{value_to_char()}}

\item{width_label}{(numeric) the width of the label box. Defaults to 60
pixels.}
}
\value{
numeric representing the with of the box in which the value will
appear
}
\description{
Estimate the width of the value box
}
\keyword{internal}
