% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_md.R
\name{file_cov_df_to_md}
\alias{file_cov_df_to_md}
\title{Transform the diff df into markdown}
\usage{
file_cov_df_to_md(diff_df, align = "rrrrc")
}
\arguments{
\item{diff_df}{(\code{tibble}) a diff df, the output of \code{\link[=derive_file_cov_df]{derive_file_cov_df()}}}

\item{align}{Column alignment: a character vector consisting of \code{'l'}
(left), \code{'c'} (center) and/or \code{'r'} (right). By default or if
\code{align = NULL}, numeric columns are right-aligned, and other columns
are left-aligned. If \code{length(align) == 1L}, the string will be
expanded to a vector of individual letters, e.g. \code{'clc'} becomes
\code{c('c', 'l', 'c')}, unless the output format is LaTeX.}
}
\value{
a character scalar containing markdown version of the diff df
collapsed into a single string.
}
\description{
Makes the column names and the contents of the \code{delta} column more readable.
Transforms the output to markdown and collapses into a single string.
}
\keyword{internal}
