% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_md.R
\name{file_cov_to_md}
\alias{file_cov_to_md}
\title{Transform the file coverage tibble into markdown}
\usage{
file_cov_to_md(file_cov_df, align = "rrrcc")
}
\arguments{
\item{file_cov_df}{a \code{tibble}, the output of \code{\link[=combine_file_coverage]{combine_file_coverage()}}.}

\item{align}{Column alignment: a character vector consisting of \code{'l'}
(left), \code{'c'} (center) and/or \code{'r'} (right). By default or if
\code{align = NULL}, numeric columns are right-aligned, and other columns
are left-aligned. If \code{length(align) == 1L}, the string will be
expanded to a vector of individual letters, e.g. \code{'clc'} becomes
\code{c('c', 'l', 'c')}, unless the output format is LaTeX.}
}
\value{
a character scalar containing markdown version of the diff df
collapsed into a single string.
}
\description{
Makes the column names human readable, formats the data, and adds an
interpretation column. The transforms the output to markdown and collapses
into a single string.
}
\keyword{internal}
