% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_coverage.R
\name{file_coverage}
\alias{file_coverage}
\title{Digest file coverage}
\usage{
file_coverage(x)
}
\arguments{
\item{x}{\verb{<coverage>} object, defaults to \code{\link[covr:package_coverage]{covr::package_coverage()}}.}
}
\value{
a \code{tibble} with 2 columns (\code{File} and \code{Coverage}) summarising
testing coverage at file level.
}
\description{
Take a \code{coverage} object (the output of \code{\link[covr:package_coverage]{covr::package_coverage()}}, extract
the filecoverage component and transform it into a \code{data.frame}/\code{tibble}.
}
\examples{
\dontrun{
library(covr)

covr::package_coverage("myawesomepkg") |>
  file_coverage()
}
}
