% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge.R
\name{generate_badge}
\alias{generate_badge}
\title{Generate a coverage badge}
\usage{
generate_badge(value)
}
\arguments{
\item{value}{(a numeric scalar) percentage coverage.}
}
\value{
a SVG string invisibly
}
\description{
Generate a coverage badge
}
\examples{
generate_badge(5)
generate_badge(48)
}
