% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_details.R
\name{get_changed_files}
\alias{get_changed_files}
\title{Get changed files}
\usage{
get_changed_files(
  owner = character(),
  repo = character(),
  pr_number = integer(),
  call = rlang::caller_env()
)
}
\arguments{
\item{owner}{(character) the repo owner.}

\item{repo}{(character) the repo name}

\item{pr_number}{(integer) the PR number}

\item{call}{the execution environment to surface the error message from.
Defaults to \code{\link[rlang:stack]{rlang::caller_env()}}.}
}
\value{
a character vector containing the names of the changed files.
}
\description{
Sends a GET request to the GitHub API and retrieves the relevant files
involved in the PR. It only includes files that are under \verb{R/} or \verb{src/}.
}
\examples{
\dontrun{
get_changed_files("dragosmg", "covr2mddemo", 2)
}
}
