% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment.R
\name{get_comment_id}
\alias{get_comment_id}
\title{Get the ID of the covr2md PR comment (if it exists)}
\usage{
get_comment_id(
  repo,
  pr_number,
  marker = "<!-- covr2md-code-coverage -->",
  call = rlang::caller_env()
)
}
\arguments{
\item{repo}{(character) the repository name in the GitHub format
(\code{"OWNER/REPO"}).}

\item{pr_number}{(integer) the PR number}

\item{marker}{(character scalar) string used to identify an issue
comment generated with covr2md. Defaults to
\code{"<!-- covr2md-code-coverage -->"}.}

\item{call}{the execution environment to surface the error message from.
Defaults to \code{\link[rlang:stack]{rlang::caller_env()}}.}
}
\value{
a numeric scalar representing the comment id
}
\description{
These are identified by the \code{"<!-- covr2md-code-coverage -->"} marker. We
need the ID of the comment before updating it.
}
\examples{
\dontrun{
get_comment_id("dragosmg/demo-repo", 3)
}
}
\keyword{internal}
