% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_details.R
\name{get_diff_line_coverage}
\alias{get_diff_line_coverage}
\title{Get the line coverage for the diff}
\usage{
get_diff_line_coverage(repo, pr_details, relevant_files, head_coverage)
}
\arguments{
\item{repo}{(character) the repository name in the GitHub format
(\code{"OWNER/REPO"}).}

\item{pr_details}{a \verb{<pr_details>} object representing a subset of the pull
request metadata we need. The output of \code{\link[=get_pr_details]{get_pr_details()}}.}

\item{relevant_files}{(character) files with changes in coverage}

\item{head_coverage}{(coverage) active / current branch (\code{HEAD}) coverage.
The output of \code{\link[covr:package_coverage]{covr::package_coverage()}} on the branch.}
}
\value{
a \code{tibble} with 3 columns:
\itemize{
\item file: file name
\item lines_added: total number of lines that would be added by merging the PR
\item lines_covered: number of added lines covered by unit tests
}
}
\description{
Are the added lines covered by unit tests?
Does this in several steps:
\itemize{
\item get the text of the git diff (the combined diff format)
\item extracts the added lines and calculates the new line numbers
\item does a bit of shuffling of the head coverage data to summarise at
line level
\item summarises the number of lines added and number of lines covered by
tests at file level
}
}
\keyword{internal}
