% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_details.R
\name{get_pr_details}
\alias{get_pr_details}
\title{Get pull request details}
\usage{
get_pr_details(repo, pr_number, call = rlang::caller_env())
}
\arguments{
\item{repo}{(character) the repository name in the GitHub format
(\code{"OWNER/REPO"}).}

\item{pr_number}{(integer) the PR number}

\item{call}{the execution environment to surface the error message from.
Defaults to \code{\link[rlang:stack]{rlang::caller_env()}}.}
}
\value{
an object of class \verb{<pr_details>} - a list with the following
elements:
\itemize{
\item \code{head_name}: name of the current branch
\item \code{head_sha}: the sha of the last commit in the current branch
\item \code{base_name}: the name of the destination branch
\item \code{base_sha}: the sha of most recent commit on the destination branch
\item \code{pr_html_url}: the URL to the PR HTML branch
}
}
\description{
Sends a GET request to the GitHub API and fetches the PR details. The output
contains a subset of these, needed for downstream use.
}
\examples{
\dontrun{
get_pr_details("dragosmg/covr2mddemo", 2)
}
}
