% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_md.R
\name{line_cov_to_md}
\alias{line_cov_to_md}
\title{Transform a line coverage tibble to markdown}
\usage{
line_cov_to_md(diff_line_coverage, align = "rrrr")
}
\arguments{
\item{diff_line_coverage}{(\code{tibble}) diff line coverage data. The output of
\code{\link[=derive_file_cov_df]{derive_file_cov_df()}}}

\item{align}{Column alignment: a character vector consisting of \code{'l'}
(left), \code{'c'} (center) and/or \code{'r'} (right). By default or if
\code{align = NULL}, numeric columns are right-aligned, and other columns
are left-aligned. If \code{length(align) == 1L}, the string will be
expanded to a vector of individual letters, e.g. \code{'clc'} becomes
\code{c('c', 'l', 'c')}, unless the output format is LaTeX.}
}
\value{
a markdown table as a string
}
\description{
Adds a total row, make column names more readable. Then transforms it to
markdown which gets collapsed into a single string.
}
\keyword{internal}
