% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment.R
\name{post_comment}
\alias{post_comment}
\title{Post or update comment}
\usage{
post_comment(body, repo, pr_number, new = FALSE, delete = new)
}
\arguments{
\item{body}{(character scalar) the content of the body of the message.}

\item{repo}{(character) the repository name in the GitHub format
(\code{"OWNER/REPO"}).}

\item{pr_number}{(integer) the PR number}

\item{new}{(logical) post a new comment or update existing one. Defaults to
\code{FALSE}.}

\item{delete}{(logical) whether to delete a comment. Useful when posting new
comments (delete old ones).}
}
\value{
a \code{gh_response} object containing the API response
}
\description{
\code{post_comment()} first checks if a "known" \code{covr2gh} comment exists on the
target pull request. If it does, then it updates it, if it doesn't, then a
a new comment is posted.
}
\examples{
\dontrun{
post_comment(
  "this is amazing",
  repo = "<owner>/<repo>",
  pr_number = 3
)
}
}
