% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_action.R
\name{use_covr2gh_badge}
\alias{use_covr2gh_badge}
\title{Add a coverage README badge}
\usage{
use_covr2gh_badge(repo)
}
\arguments{
\item{repo}{(character) the repository name in the GitHub format
(\code{"OWNER/REPO"}).}
}
\description{
Inserts the markdown text for a coverage badge in the usethis block (between
\verb{<!-- badges: start -->} and \verb{<!-- badges: end -->}). The \code{src} URL for the
badge SVG points to a storage branch in the repo(\code{covr2gh-storage}). The link
out points to the actions workflow page.
}
\examples{
\dontrun{
use_covr2gh_badge("<owner>/<repo>")
}
}
