% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge_params.R
\name{value_to_char}
\alias{value_to_char}
\title{Transform a coverage value into char}
\usage{
value_to_char(value, verbose = FALSE, call = rlang::caller_env())
}
\arguments{
\item{value}{(numeric) coverage value. Can also be \code{NA} or \code{NULL}}

\item{verbose}{(logical) if we want messages around the clamping. Default is
\code{FALSE}.}

\item{call}{(call) defaults to \code{\link[rlang:stack]{rlang::caller_env()}}}
}
\value{
a list with 2 element:
\itemize{
\item \code{num}: the value as numeric
\item \code{char}: the value as character
}
}
\description{
Outputs both values as both are used downstream by various other helpers.
If the input values are \code{NA} or \code{NULL} it converts them to \code{"unknown"}.
If the values are greater than 100 or less than 0 it adjusts them to these
values
}
\examples{
\dontrun{
value_to_char(26.78)
}
}
\keyword{internal}
