% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose.R
\name{compose_comment}
\alias{compose_comment}
\title{Compose a coverage comment}
\usage{
compose_comment(
  head_coverage,
  base_coverage,
  repo,
  pr_number,
  diff_cov_target = NULL
)
}
\arguments{
\item{head_coverage}{(coverage) active / current branch (\code{HEAD}) coverage.
The output of \code{\link[covr:package_coverage]{covr::package_coverage()}} on the head branch.}

\item{base_coverage}{(coverage) base / target branch coverage (coverage for
the branch merging into). The output of \code{\link[covr:package_coverage]{covr::package_coverage()}} on the
base branch.}

\item{repo}{(character) the repository name in the GitHub format
(\code{"OWNER/REPO"}).}

\item{pr_number}{(integer) the PR number}

\item{diff_cov_target}{(numeric) minimum accepted diff coverage. Defaults to
\code{NULL} which is then interpreted as overall base branch coverage.}
}
\value{
a character scalar with the content of the GitHub comment
}
\description{
Compose a coverage comment
}
\examples{
\dontrun{
coverage_head <- covr::package_coverage()
system("git checkout main")
coverage_main <- covr::package_coverage()

compose_comment(
  head_coverage = coverage_head,
  base_coverage = coverage_main,
  repo = "<owner>/<repo>",
  pr_number = 3
)
}
}
