% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment.R
\name{delete_comment}
\alias{delete_comment}
\title{Delete a comment}
\usage{
delete_comment(owner, repo, comment_id)
}
\arguments{
\item{owner}{(character) the repo owner.}

\item{repo}{(character) the repo name}

\item{comment_id}{(numeric) the ID of the issue comment to delete.}
}
\value{
a \code{gh_response} object
}
\description{
Thin wrapper for making a \code{DELETE} request to the GitHub API.
}
\examples{
\dontrun{
delete_comment("dragosmg", "this-is-my-repo", 4553)
}
}
\keyword{internal}
