% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_details.R
\name{get_changed_files}
\alias{get_changed_files}
\title{Get changed files}
\usage{
get_changed_files(repo, pr_number, call = rlang::caller_env())
}
\arguments{
\item{repo}{(character) the repository name in the GitHub format
(\code{"OWNER/REPO"}).}

\item{pr_number}{(integer) the PR number}

\item{call}{the execution environment to surface the error message from.
Defaults to \code{\link[rlang:stack]{rlang::caller_env()}}.}
}
\value{
a character vector containing the names of the changed files.
}
\description{
Sends a GET request to the GitHub API and retrieves the files modified by
the PR. It then subsets these to only includes the "relevant" files - i.e.
those under \verb{R/} or \verb{src/}.
}
\examples{
\dontrun{
get_changed_files("<owner>/<repo>", 2)
}
}
\keyword{internal}
