% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment.R
\name{get_comment_id}
\alias{get_comment_id}
\title{Identify the covr2gh comment}
\usage{
get_comment_id(repo, pr_number, call = rlang::caller_env())
}
\arguments{
\item{repo}{(character) the repository name in the GitHub format
(\code{"OWNER/REPO"}).}

\item{pr_number}{(integer) the PR number}

\item{call}{the execution environment to surface the error message from.
Defaults to \code{\link[rlang:stack]{rlang::caller_env()}}.}
}
\value{
a numeric scalar representing the comment id or \code{NULL}
}
\description{
Comments posted by covr2gh are identified by the presence of the
\code{"<!-- covr2gh-do-not-delete -->"} comment. \code{get_comment_id()} looks for it.
If it can find it, it returns the comment ID, otherwise it returns \code{NULL}.
}
\details{
The output of \code{get_comment_id()} is then used \code{post_comment()} post a new
comment or update an existing one.
}
\examples{
\dontrun{
get_comment_id("<owner>/<repo>", 3)
}
}
\keyword{internal}
