% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment.R
\name{get_comment_id}
\alias{get_comment_id}
\title{Get the ID of the covr2md comment}
\usage{
get_comment_id(
  repo,
  pr_number,
  marker = "<!-- covr2md-code-coverage -->",
  call = rlang::caller_env()
)
}
\arguments{
\item{repo}{(character) the repository name in the GitHub format
(\code{"OWNER/REPO"}).}

\item{pr_number}{(integer) the PR number}

\item{marker}{(character scalar) string used to identify an issue
comment generated with covr2md. Defaults to
\code{"<!-- covr2md-code-coverage -->"}.}

\item{call}{the execution environment to surface the error message from.
Defaults to \code{\link[rlang:stack]{rlang::caller_env()}}.}
}
\value{
a numeric scalar representing the comment id or \code{NULL}
}
\description{
Comments are identified by a specific comment, the "marker", by default
\code{"<!-- covr2md-code-coverage -->"}. \code{get_comment_id()} looks for this
marker. If it can find it, it returns the comment ID, otherwise it returns
\code{NULL}.
}
\details{
The output of \code{get_comment_id()} is the used downstream by \code{post_comment()}
post a new comment or to update an existing one.
}
\examples{
\dontrun{
get_comment_id("dragosmg/demo-repo", 3)
}
}
