% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_details.R
\name{get_diff_text}
\alias{get_diff_text}
\title{Get the PR diff}
\usage{
get_diff_text(repo, pr_details, relevant_files, call = rlang::caller_env())
}
\arguments{
\item{repo}{(character) the repository name in the GitHub format
(\code{"OWNER/REPO"}).}

\item{call}{the execution environment to surface the error message from.
Defaults to \code{\link[rlang:stack]{rlang::caller_env()}}.}
}
\value{
a character vector containing the names of the changed files.
}
\description{
Sends a GET request to the GitHub API and retrieves the files modified by
the PR. It then subsets these to only includes files under \verb{R/} or \verb{src/}.
}
\examples{
\dontrun{
get_diff_text("<owner>/<repo>", 2)
}
}
\keyword{internal}
