% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_details.R
\name{get_diff_text}
\alias{get_diff_text}
\title{Get the PR diff}
\usage{
get_diff_text(pr_details, relevant_files, call = rlang::caller_env())
}
\arguments{
\item{pr_details}{a \verb{<pr_details>} object representing a subset of the pull
request metadata we need. The output of \code{\link[=get_pr_details]{get_pr_details()}}.}

\item{relevant_files}{(character) files with changes in coverage}

\item{call}{the execution environment to surface the error message from.
Defaults to \code{\link[rlang:stack]{rlang::caller_env()}}.}
}
\value{
a character vector containing the names of the changed files.
}
\description{
Sends a GET request to the GitHub API and retrieves the files modified by
the PR. It then subsets these to only includes files under \verb{R/} or \verb{src/}.
}
\examples{
\dontrun{
pr_details <- get_pr_details("<owner>/<repo>", 2)
# TODO example
relevant_files <-
diff_text <- get_diff_text(pr_details, relevant files)
}
}
\keyword{internal}
