% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pr_details.R
\name{get_pr_details}
\alias{get_pr_details}
\title{Get pull request details}
\usage{
get_pr_details(
  owner = character(),
  repo = character(),
  pr_number = integer(),
  call = rlang::caller_env()
)
}
\arguments{
\item{owner}{(character) the repo owner.}

\item{repo}{(character) the repo name}

\item{pr_number}{(integer) the PR number}

\item{call}{the execution environment to surface the error message from.
Defaults to \code{\link[rlang:stack]{rlang::caller_env()}}.}
}
\value{
a list with the following elements:
\itemize{
\item \code{head_name}: name of the current branch
\item \code{head_sha}: the sha of the last commit in the current branch
\item \code{base_name}: the name of the destination branch
\item \code{base_sha}: the sha of most recent commit on the destination branch
\item \code{pr_html_url}: the URL to the PR HTML branch
}
}
\description{
Sends a GET request to the GitHub API and outputs details about the PR.
}
\examples{
\dontrun{
get_pr_details("dragosmg", "covr2mddemo", "2")
}
}
