% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comment.R
\name{post_comment}
\alias{post_comment}
\title{Post a new comment or update an existing one}
\usage{
post_comment(
  body,
  repo,
  pr_number,
  comment_id = NULL,
  marker = "<!-- covr2md-code-coverage -->"
)
}
\arguments{
\item{body}{(character scalar) the content of the body of the message.}

\item{repo}{(character) the repository name in the GitHub format
(\code{"OWNER/REPO"}).}

\item{pr_number}{(integer) the PR number}

\item{comment_id}{(numeric) the ID of the issue comment to update. If \code{NULL}
(the default), a new comment will be posted.}

\item{marker}{(character scalar) string used to identify an issue
comment generated with covr2md. Defaults to
\code{"<!-- covr2md-code-coverage -->"}.}
}
\value{
a \code{gh_response} object containing the API response
}
\description{
Post a new comment or update an existing one
}
\examples{
\dontrun{
post_comment(
  "this is amazing",
  repo = "dragosmg/my-test-repo",
  pr_number = 3,
  comment_id = 123456789,
  marker = "<!-- I am a comment -->"
)
}
}
