% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_action.R
\name{use_covr2gh_action}
\alias{use_covr2gh_action}
\title{Set up a covr2gh GitHub Action workflow}
\usage{
use_covr2gh_action(badge = TRUE)
}
\arguments{
\item{badge}{(logical) should a badge be added to README? Defaults to \code{TRUE}.}
}
\description{
Sets up a \href{https://github.com/features/actions}{GitHub Actions} workflow that
calculates and reports test coverage:
\itemize{
\item uses \code{\link[covr:package_coverage]{covr::package_coverage()}}.
\item on a \emph{pull request}:
\itemize{
\item calculates test coverage for head.
\item builds a badge (for use in the commit message).
\itemize{
\item if the PR does not come from a fork (write privileges for the
automated GitHub Actions token), the badge is built with an internal
function (\code{\link[=generate_badge]{generate_badge()}}) and committed to the \code{covr2gh-storage}
branch.
\item if the PR comes from a fork (the GitHub Actions token only has read
permissions), the badge is built with an external service (shields.io)
and is not committed to the storage branch.
}
\item initialises or switches to the \code{covr2gh-storage} branch and saves a
copy of the badge (only if the PR does not come from a fork).
\item switches to base branch and calculates test coverage.
\item compares coverage for head and base and posts a comment with the main
findings.
\item uploads base coverage and badge as artifacts.
\item posts a workflow \href{https://github.blog/news-insights/product-news/supercharging-github-actions-with-job-summaries/}{summary}.
}
\item on a \emph{push to main}:
\itemize{
\item calculates test coverage for head and creates a badge.
\item switches to or creates the \code{covr2gh-storage} branch.
\item commits badge (which is referenced in the README).
\item uploads coverage as an artefact.
}
}
}
\details{
\code{use_covr2gh_action()} wraps \code{\link[usethis:use_github_action]{usethis::use_github_action()}}.
}
\examples{
\dontrun{
use_covr2gh_action()
}
}
