% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_action.R
\name{use_covr2gh_badge}
\alias{use_covr2gh_badge}
\title{Add a coverage README badge}
\usage{
use_covr2gh_badge()
}
\description{
Inserts the markdown text for a coverage badge in the usethis block (between
\verb{<!-- badges: start -->} and \verb{<!-- badges: end -->}). The URL for the badge
SVG points to a storage branch in the repo(\code{covr2gh-storage}). The link out
points to the covr2gh pkgdown website (not yet - TODO update once there is
an article to point to).
}
\examples{
\dontrun{
use_covr2gh_badge()
}
}
